pyproject.toml
[project]
name = "ai-security-agent-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "flask==3.1.1",
    "flask-cors==6.0.0",
    "Flask-SQLAlchemy==3.1.1",
    "uvicorn==0.35.0",
    "uvloop==0.21.0",
    "watchfiles==1.1.0",

    # Config and utils
    "python-dotenv==1.1.1",
    "requests==2.32.5",
    "pydantic==2.11.7",
    "tenacity==9.1.2",
    "authlib>=1.3.0",

    # Azure SDKs
    "azure-identity>=1.17.1",
    "azure-core>=1.30.2",
    "azure-storage-blob>=12.23.1",
    "azure-search-documents>=11.6.0b8",
    "azure-ai-documentintelligence>=1.0.0",

    # Azure OpenAI via official client
    "openai==1.101.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
    "pytest>=7.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
]
